ci: format: add workflow that checks formatting of all sub-projects
Discussing personal formatting preferences is a tedious topic.
Luckily it is mostly a topic of the past, because automatic formatters
exists.
They may not always produce the _best_ result, but they do produce a good
and uniform result and they save us from ever having to discuss formatting
in pull requests again.

Run the default dart formatter in the CI to ensure new code is formatted
correctly. Also re-check the formatting periodically to check if the
correct formatting has changed across dart versions.